Using Boost Unit Test properly, run all tests
Before this change I had not taken the time to look into how to properly set up the Boost
Unit Test Framework to compile from multiple test files into a single binary. There is still
a lot of clean up to do around the build, as there are no build scripts, but that will come
in time.